Sample fails with 0.0.6 release of azure-event-hubs node module, throttling back to 0.0.3 for now.
Giselle has isolated the azure-event-hubs plugin as having some issues with this sample. She is trying to isolate the reason for the problem. When she discovers what that is, we can update this library version and/or the sample. Until then, this node module needs to remain fixed to a specific version that we know works with the sample!